move checkModelAvailability into textInference store and remove provide/inject
