+++ WebUI/src/assets/js/store/textInference.ts
@@ -5,6 +5,7 @@ import { useModels } from './models'
 import * as Const from '@/assets/js/const'
 import { Document } from 'langchain/document'
 import { llmBackendTypes } from '@/types/shared'
+import { useDialogStore } from "@/assets/js/store/dialogs.ts";
 
 const LlmBackendSchema = z.enum(llmBackendTypes)
 export type LlmBackend = z.infer<typeof LlmBackendSchema>
@@ -77,16 +78,17 @@ export const textInferenceBackendDescription: Record<LlmBackend, string> = {
 }
 
 export const textInferenceBackendTags: Record<LlmBackend, string[]> = {
-  ipexLLM: ['Intel','Legacy'],
-  llamaCPP: ['Lightweight','Portable'],
-  openVINO: ['Intel','Optimized','Fast'],
-  ollama: ['Integrated','CLI'],
+  ipexLLM: ['Intel', 'Legacy'],
+  llamaCPP: ['Lightweight', 'Portable'],
+  openVINO: ['Intel', 'Optimized', 'Fast'],
+  ollama: ['Integrated', 'CLI'],
 }
 
 export const useTextInference = defineStore(
   'textInference',
   () => {
     const backendServices = useBackendServices()
+    const dialogStore = useDialogStore()
     const models = useModels()
     const backend = ref<LlmBackend>('openVINO')
     const ragList = ref<IndexedDocument[]>([])
@@ -357,6 +359,7 @@ export const useTextInference = defineStore(
         fontSizeIndex.value++
       }
     }
+
     function decreaseFontSize() {
       if (!isMinSize.value) {
         fontSizeIndex.value--
@@ -394,7 +397,7 @@ export const useTextInference = defineStore(
         embeddingModel: activeEmbeddingModel.value,
         maxResults: runningOnOpenvinoNpu.value ? 2 : 8,
       }
-      console.log('trying to request rag for', { newEmbedInquiry, ragList: ragList.value })
+      console.log('trying to request rag for', {newEmbedInquiry, ragList: ragList.value})
       const response = await window.electronAPI.embedInputUsingRag(newEmbedInquiry)
       return response
     }
@@ -570,7 +573,7 @@ export const useTextInference = defineStore(
           if (pageEntries.some((e) => !e.lines)) {
             // If we haven't already added an entry for this page from the line merging
             if (!result.some((r) => r.page === pageNumber && !r.lines)) {
-              result.push({ page: pageNumber })
+              result.push({page: pageNumber})
             }
           }
         })
@@ -654,6 +657,24 @@ export const useTextInference = defineStore(
       }
     }
 
+    async function checkModelAvailability() {
+      // ToDo: the path for embedding downloads must be corrected and BAAI/bge-large-zh-v1.5 was accidentally downloaded to the wrong place
+      return new Promise<void>(async (resolve, reject) => {
+        const requiredModelDownloads =
+          await getDownloadParamsForCurrentModelIfRequired('llm')
+        if (ragList.value.length > 0) {
+          const requiredEmbeddingModelDownloads =
+            await getDownloadParamsForCurrentModelIfRequired('embedding')
+          requiredModelDownloads.push(...requiredEmbeddingModelDownloads)
+        }
+        if (requiredModelDownloads.length > 0) {
+          dialogStore.showDownloadDialog(requiredModelDownloads, resolve, reject)
+        } else {
+          resolve()
+        }
+      })
+    }
+
     return {
       backend,
       activeModel,
@@ -688,6 +709,7 @@ export const useTextInference = defineStore(
       extractPostMarker,
       formatRagSources,
       ensureBackendReadiness,
+      checkModelAvailability,
 
       // Backend preparation state and methods
       isPreparingBackend: computed(() => backendReadinessState.isPreparingBackend),
