+++ WebUI/src/components/AddLLMDialog.vue
@@ -66,7 +66,6 @@ const showInfo = ref(false)
 const addModelError = ref(false)
 const animate = ref(false)
 
-const checkModelAvailability = inject<() => void>('checkModelAvailability')
 const emits = defineEmits<{
   (e: 'close'): void
 }>()
@@ -124,7 +123,7 @@ async function addModel() {
       default: false,
     })
     textInference.selectModel(textInference.backend, modelRequest.value)
-    checkModelAvailability?.()
+    await textInference.checkModelAvailability()
     closeAdd()
   }
